11:35_03/11_Estefane

+++ src/main/java/com/sisu/sisu/Dao/IGradoDao.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.GradoAcademico;
 
-public interface IGradoDao extends CrudRepository<GradoAcademico, Long> {
+public interface IGradoDao extends CrudRepository<GradoAcademico, Integer> {
 
 }
