+++ src/main/java/com/sisu/sisu/Dao/TipoRecetaDao.java
@@ -4,6 +4,6 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.sisu.sisu.entitys.TipoReceta;
 
-public interface TipoRecetaDao extends CrudRepository<TipoReceta, Long> {
+public interface TipoRecetaDao extends CrudRepository<TipoReceta, Integer> {
     
 }
