+++ src/main/java/com/sisu/sisu/controller/PersonaUsuarioController.java
@@ -73,7 +73,7 @@ public class PersonaUsuarioController {
 	public String guardarPersona(@Validated Persona persona1, RedirectAttributes flash,HttpServletRequest request,
     @RequestParam(name="apodo",required = false)String apodo,
     @RequestParam(name="clave",required = false)String clave,
-    @RequestParam(name="grado",required = false)Long idGradoAcademico,
+    @RequestParam(name="grado",required = false)Integer idGradoAcademico,
     @RequestParam(name="dip",required = false)Long idDip,
     @RequestParam(name="estadoCivil",required = false)Long idTipoEstadoCivil
     ) { 
@@ -108,7 +108,7 @@ public class PersonaUsuarioController {
         @RequestParam(name = "Direccion") String Direccion,
         @RequestParam(name = "apodo") String apodo,
         @RequestParam(name = "clave") String clave,
-        @RequestParam(name="grado",required = false)Long idGradoAcademico,
+        @RequestParam(name="grado",required = false)Integer idGradoAcademico,
         @RequestParam(name="dip",required = false)Long idDip,
         @RequestParam(name="estadoCivil",required = false)Long idTipoEstadoCivil
         ) {
